+++ README.md
@@ -1,11 +1,11 @@
 | Check | TODO                                 |
 | :---: | ------------------------------------ |
 |  ✅   | 일기 수정,일기 삭제                  |
-|  ⬜️  | 내정보 수정하기                      |
 |  ✅   | 좋아요 기능                          |
-|  ⬜️  | 날씨 선택기능                        |
 |  ✅   | OG태그                               |
-|  ⬜️  | 지역 추가 및 선택 할 수 있도록       |
 |  ✅   | 일기 쓸 때 위도 경도 선택            |
+|  ⬜️  | 내정보 수정하기                      |
+|  ⬜️  | 날씨 선택기능                        |
+|  ⬜️  | 지역 추가 및 선택 할 수 있도록       |
 |  ⬜️  | 화면 무한스크롤 해서 첫 페이지는 ssr |
 |  ⬜️  | SEO 작업                             |
