+++ src/components/common/fontPre.tsx
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+const FontPre = styled.pre`
+  font-family: Chilgok_lws;
+  white-space: pre-wrap;
+`;
+
+export default FontPre;
